Workwise_V3

+++ src/api-gateway/src/company/company.controller.ts
@@ -11,6 +11,8 @@ import {
   HttpStatus,
   Query,
   Headers,
+  UseInterceptors,
+  UploadedFile,
 } from '@nestjs/common';
 import { CompanyService } from './company.service';
 import {
@@ -19,11 +21,17 @@ import {
   CreateCompanyResponseDto,
   findCompanyResponseDto,
 } from './dto/create-company.dto';
-import { UpdateCompanyDto } from './dto/update-company.dto';
+import {
+  UpdateCompanyDto,
+  UpdateCompanyJobStatusesDto,
+  UpdateCompanyLogoDto,
+  UpdateCompanyJobStatuses,
+} from './dto/update-company.dto';
 import { AuthGuard } from '../auth/auth.guard';
 import {
   ApiBearerAuth,
   ApiBody,
+  ApiConsumes,
   ApiInternalServerErrorResponse,
   ApiOkResponse,
   ApiOperation,
@@ -37,13 +45,16 @@ import mongoose, { FlattenMaps, Types } from 'mongoose';
 import {
   Company,
   CompanyAllResponseDto,
-  CompanyEmployeesResponseDto,
+  CompanyDetailedResponseDto,
   CompanyResponseDto,
 } from './entities/company.entity';
 import { DeleteEmployeeFromCompanyDto } from './dto/delete-employee-in-company.dto';
-import { validateObjectIds } from '../utils/Utils';
+import { extractUserId, validateObjectId, validateObjectIds } from '../utils/Utils';
 import { JwtService } from '@nestjs/jwt';
 import { BooleanResponseDto } from '../shared/dtos/api-response.dto';
+import { FileInterceptor } from '@nestjs/platform-express';
+import { JobStatusAllResponseDto } from '../job/dto/job-responses.dto';
+import { EmployeeService } from '../employee/employee.service';
 
 const className = 'Company';
 
@@ -53,8 +64,8 @@ export class CompanyController {
   constructor(
     private readonly companyService: CompanyService,
     private readonly jwtService: JwtService,
+    private readonly employeeService: EmployeeService,
   ) {}
-
   validateObjectId(id: string | Types.ObjectId): boolean {
     if (!mongoose.Types.ObjectId.isValid(id)) {
       throw new HttpException('Invalid ID', HttpStatus.BAD_REQUEST);
@@ -83,9 +94,7 @@ export class CompanyController {
     description: `The access token and ${className}'s Id used for querying.`,
   })
   @Post('/create')
-  async create(
-    @Body() createCompanyDto: CreateCompanyDto,
-  ): Promise<CreateCompanyResponseDto> {
+  async create(@Body() createCompanyDto: CreateCompanyDto): Promise<CreateCompanyResponseDto> {
     return await this.companyService.create(createCompanyDto);
   }
 
@@ -97,6 +106,11 @@ export class CompanyController {
   @ApiOkResponse({ type: BooleanResponseDto })
   @Post('/add')
   async addEmployee(@Body() addUserDto: AddUserToCompanyDto) {
+    // const currentEmployee = await this.employeeService.findById(
+    //   body.currentEmployeeId,
+    // );
+    // if (currentEmployee.role.permissionSuite.includes('add new employees')) {
+    //TODO: Figure out if this endpoint needs role based access
     const arr = [addUserDto.adminId, addUserDto.currentCompany];
     if (addUserDto.roleId) arr.push(addUserDto.roleId);
     validateObjectIds(arr);
@@ -106,6 +120,7 @@ export class CompanyController {
     } catch (Error) {
       throw new HttpException('Internal server error', HttpStatus.CONFLICT);
     }
+    // }
   }
 
   @UseGuards(AuthGuard) //It may be accessed by external users
@@ -121,10 +136,7 @@ export class CompanyController {
     try {
       return { data: await this.companyService.getAllCompanyNames() };
     } catch (Error) {
-      throw new HttpException(
-        'Something went wrong',
-        HttpStatus.INTERNAL_SERVER_ERROR,
-      );
+      throw new HttpException('Something went wrong', HttpStatus.INTERNAL_SERVER_ERROR);
     }
   }
 
@@ -141,34 +153,7 @@ export class CompanyController {
     try {
       return { data: await this.companyService.getAllCompanies() };
     } catch (Error) {
-      throw new HttpException(
-        'Something went wrong',
-        HttpStatus.INTERNAL_SERVER_ERROR,
-      );
-    }
-  }
-
-  @UseGuards(AuthGuard) //Need to add authorization
-  @ApiOperation({ summary: `Get all employees in ${className}` })
-  @ApiOkResponse({
-    type: CompanyEmployeesResponseDto,
-    description: `The mongodb object of the ${className}, with an _id attribute`,
-  })
-  @ApiParam({
-    name: 'cid',
-    description: `The _id attribute of the ${className}`,
-  })
-  @Get('/all/employees/:cid')
-  async getAllEmployeesInCompany(@Param('cid') cid: string) {
-    this.validateObjectId(cid);
-    const objId = new Types.ObjectId(cid);
-    try {
-      return { data: await this.companyService.getAllEmployees(objId) };
-    } catch (Error) {
-      throw new HttpException(
-        'Something went wrong',
-        HttpStatus.INTERNAL_SERVER_ERROR,
-      );
+      throw new HttpException('Something went wrong', HttpStatus.INTERNAL_SERVER_ERROR);
     }
   }
 
@@ -180,9 +165,28 @@ export class CompanyController {
   @Get('id/:id')
   async findOne(@Param('id') id: string) {
     try {
-      this.validateObjectId(id);
+      validateObjectId(id);
       return {
         data: await this.companyService.getCompanyById(new Types.ObjectId(id)),
+      };
+    } catch (e) {
+      throw new HttpException(e, HttpStatus.NOT_FOUND);
+    }
+  }
+
+  @ApiOperation({
+    summary: `Find a ${className}, with Actual Employees and Inventory Items instead of ObjectIds`,
+  })
+  @ApiOkResponse({
+    type: CompanyDetailedResponseDto,
+    description: `The mongodb 'Detailed' object of the ${className}, with an _id attribute`,
+  })
+  @Get('id/:id/detailed')
+  async findOneDetailed(@Param('id') id: string): Promise<{ data: FlattenMaps<Company> & { _id: Types.ObjectId } }> {
+    try {
+      validateObjectId(id);
+      return {
+        data: await this.companyService.getCompanyByIdDetailed(new Types.ObjectId(id)),
       };
     } catch (e) {
       throw new HttpException(e, HttpStatus.NOT_FOUND);
@@ -223,30 +227,67 @@ export class CompanyController {
     description: `The updated ${className} object`,
   })
   @ApiBody({ type: UpdateCompanyDto })
-  @Patch(':cid')
+  @Patch('update/:cid')
   async update(
     @Headers() headers: any,
     @Param('cid') cid: string,
-    @Body() updateCompanyDto: UpdateCompanyDto,
+    @Body()
+    body: {
+      currentEmployeeId: Types.ObjectId;
+      updateCompanyDto: UpdateCompanyDto;
+    },
   ) {
-    try {
-      this.validateObjectId(cid);
-      const userId = this.extractUserId(headers);
+    const currentEmployee = await this.employeeService.findById(body.currentEmployeeId);
+    if (currentEmployee.role.permissionSuite.includes('company settings')) {
+      try {
+        validateObjectId(cid);
+        const userId = extractUserId(this.jwtService, headers);
 
-      const companyId = new Types.ObjectId(cid);
-      const updatedCompany = await this.companyService.update(
-        userId,
-        companyId,
-        updateCompanyDto,
-      );
-      return {
-        data: updatedCompany,
-      };
-    } catch (e) {
-      throw new HttpException(
-        'internal server error',
-        HttpStatus.INTERNAL_SERVER_ERROR,
-      );
+        const companyId = new Types.ObjectId(cid);
+        const updatedCompany = await this.companyService.update(userId, companyId, body.updateCompanyDto);
+        return {
+          data: updatedCompany,
+        };
+      } catch (e) {
+        throw new HttpException('internal server error', HttpStatus.INTERNAL_SERVER_ERROR);
+      }
+    } else {
+      throw new HttpException('Unauthorized', HttpStatus.UNAUTHORIZED);
+    }
+  }
+
+  @UseGuards(AuthGuard)
+  @ApiBearerAuth('JWT')
+  @ApiOperation({
+    summary: `Change the Logo of a ${className}`,
+  })
+  @ApiOkResponse({
+    type: CompanyResponseDto,
+    description: `The updated ${className} instance`,
+  })
+  @ApiConsumes('multipart/form-data')
+  @ApiBody({ type: UpdateCompanyLogoDto })
+  @UseInterceptors(FileInterceptor('logo'))
+  @Patch('/update/:cid/logo')
+  async updateLogo(
+    @Headers() headers: any,
+    @Param('cid') companyId: string,
+    @UploadedFile() file: Express.Multer.File,
+    @Body() body: { currentEmployeeId: Types.ObjectId },
+  ) {
+    const currentEmployee = await this.employeeService.findById(body.currentEmployeeId);
+    if (currentEmployee.role.permissionSuite.includes('company settings')) {
+      try {
+        validateObjectId(companyId);
+        const userId = extractUserId(this.jwtService, headers);
+        return {
+          data: await this.companyService.updateLogo(userId, new Types.ObjectId(companyId), file),
+        };
+      } catch (e) {
+        throw new HttpException('internal server error', HttpStatus.INTERNAL_SERVER_ERROR);
+      }
+    } else {
+      throw new HttpException('Unauthorized', HttpStatus.UNAUTHORIZED);
     }
   }
 
@@ -267,18 +308,24 @@ export class CompanyController {
     description: `The _id attribute of the ${className}`,
   })
   @Delete(':cid')
-  async remove(@Headers() headers: any, @Param('cid') cid: string) {
-    try {
-      this.validateObjectId(cid);
-      const userId = this.extractUserId(headers);
-      const objectId = new Types.ObjectId(cid);
-      await this.companyService.deleteCompany(userId, objectId);
-      return { data: true };
-    } catch (e) {
-      throw new HttpException(
-        'Internal Server Error',
-        HttpStatus.SERVICE_UNAVAILABLE,
-      );
+  async remove(
+    @Headers() headers: any,
+    @Param('cid') cid: string,
+    @Body() body: { currentEmployeeId: Types.ObjectId },
+  ) {
+    const currentEmployee = await this.employeeService.findById(body.currentEmployeeId);
+    if (currentEmployee.role.permissionSuite.includes('company settings')) {
+      try {
+        validateObjectId(cid);
+        const userId = extractUserId(this.jwtService, headers);
+        const objectId = new Types.ObjectId(cid);
+        await this.companyService.deleteCompany(userId, objectId);
+        return { data: true };
+      } catch (e) {
+        throw new HttpException('Internal Server Error', HttpStatus.SERVICE_UNAVAILABLE);
+      }
+    } else {
+      throw new HttpException('Unauthorized', HttpStatus.UNAUTHORIZED);
     }
   }
 
@@ -287,7 +334,6 @@ export class CompanyController {
   @ApiOperation({
     summary: `Delete an Employee from a company using their 'Id'`,
     description: `You send the Employee _id, and then they get deleted if the id is valid.`,
-    security: [],
   })
   @ApiOkResponse({
     type: BooleanResponseDto,
@@ -295,34 +341,124 @@ export class CompanyController {
   })
   @ApiBody({
     type: DeleteEmployeeFromCompanyDto,
-    description: '',
   })
   @Delete('/emp')
   async removeEmployee(
     @Headers() headers: any,
-    @Body() deleteEmployeeDto: DeleteEmployeeFromCompanyDto,
+
+    @Query('adminId') adminId: Types.ObjectId,
+    @Query('companyId') companyId: Types.ObjectId,
+    @Query('employeeToDeleteId') employeeToDeleteId: Types.ObjectId,
   ) {
     try {
-      const userId = this.extractUserId(headers);
-      await this.companyService.deleteEmployee(userId, deleteEmployeeDto);
-      return { data: true };
+      validateObjectId(adminId, 'employee');
+      validateObjectId(companyId, 'company');
+      validateObjectId(employeeToDeleteId, 'employeeToDelete');
     } catch (e) {
-      throw new HttpException(
-        'Internal Server Error',
-        HttpStatus.SERVICE_UNAVAILABLE,
-      );
+      throw new HttpException('Invalid request', HttpStatus.BAD_REQUEST);
+    }
+    const deleteEmployeeDto: DeleteEmployeeFromCompanyDto = new DeleteEmployeeFromCompanyDto(
+      adminId,
+      companyId,
+      employeeToDeleteId,
+    );
+    //await this.companyService.deleteEmployee(userId, deleteEmployeeDto); //TODO: Use with cascading delete
+    const userId = extractUserId(this.jwtService, headers);
+    const currentEmployee = await this.employeeService.findById(deleteEmployeeDto.adminId);
+    if (currentEmployee.role.permissionSuite.includes('remove any employees')) {
+      let data;
+      try {
+        data = await this.employeeService.remove(deleteEmployeeDto.employeeToDeleteId);
+      } catch (e) {
+        throw new HttpException('Invalid request', HttpStatus.BAD_REQUEST);
+      }
+
+      if (data === false) {
+        throw new HttpException('update unsuccessful', HttpStatus.INTERNAL_SERVER_ERROR);
+      }
+      return { data: data };
+    } else if (currentEmployee.role.permissionSuite.includes('remove employees under me')) {
+      let data;
+      try {
+        data = await this.employeeService.removeUnderMe(
+          userId,
+          deleteEmployeeDto.employeeToDeleteId,
+          deleteEmployeeDto.adminId,
+        );
+      } catch (e) {
+        throw new HttpException('Invalid request', HttpStatus.BAD_REQUEST);
+      }
+
+      if (data === false) {
+        throw new HttpException('update unsuccessful', HttpStatus.INTERNAL_SERVER_ERROR);
+      }
+      return { data: data };
+    } else {
+      throw new HttpException('Unauthorized', HttpStatus.UNAUTHORIZED);
     }
   }
 
-  private extractUserId(headers: any) {
-    const authHeader: string = headers.authorization;
-    const decodedJwtAccessToken = this.jwtService.decode(
-      authHeader.replace(/^Bearer\s+/i, ''),
-    );
-    if (!Types.ObjectId.isValid(decodedJwtAccessToken.sub)) {
-      throw new HttpException('Invalid User', HttpStatus.BAD_REQUEST);
+  @UseGuards(AuthGuard)
+  @ApiBearerAuth('JWT')
+  @ApiOperation({
+    summary: `Get all Statuses in a company`,
+  })
+  @ApiOkResponse({
+    type: JobStatusAllResponseDto,
+    description: `An array of JobsStatuses`,
+  })
+  @Get('status/all/:cid')
+  async findAllStatusInCompany(@Headers() headers: any, @Param('cid') cId: string) {
+    try {
+      validateObjectId(cId);
+      const companyId = new Types.ObjectId(cId);
+      const userId: Types.ObjectId = extractUserId(this.jwtService, headers);
+      return {
+        data: await this.companyService.findAllStatusesInCompany(userId, companyId),
+      };
+    } catch (e) {
+      console.log(e);
+      throw e;
     }
-    const userId: Types.ObjectId = decodedJwtAccessToken.sub; //This attribute is retrieved in the JWT
-    return userId;
+  }
+
+  @UseGuards(AuthGuard)
+  @ApiBearerAuth('JWT')
+  @ApiOperation({
+    summary: `Update the order of Job Statuses in company, this will be used for the kanban`,
+    description: '',
+  })
+  @ApiOkResponse({
+    type: JobStatusAllResponseDto,
+    description: `The updated array`,
+  })
+  @ApiBody({ type: UpdateCompanyJobStatusesDto })
+  @Patch('statuses')
+  async updateStatusOrder(
+    @Headers() headers: any,
+    @Body()
+    body: {
+      currentEmployeeId: Types.ObjectId;
+      updateCompanyJobStatusesDto: UpdateCompanyJobStatusesDto;
+    },
+  ) {
+    const currentEmployee = await this.employeeService.findById(body.currentEmployeeId);
+    if (currentEmployee.role.permissionSuite.includes('company settings')) {
+      try {
+        const userId = extractUserId(this.jwtService, headers);
+        const statusArr = new UpdateCompanyJobStatuses(body.updateCompanyJobStatusesDto);
+        return {
+          data: await this.companyService.updateCompanyStatuses(
+            userId,
+            body.updateCompanyJobStatusesDto.employeeId,
+            statusArr,
+          ),
+        };
+      } catch (e) {
+        throw e;
+      }
+    } else {
+      throw new HttpException('Unauthorized', HttpStatus.UNAUTHORIZED);
+    }
   }
 }
